+++ files/bootstrap.sh
@@ -23,10 +23,11 @@ dnf install git -y
 dnf config-manager --add-repo https://cli.github.com/packages/rpm/gh-cli.repo
 dnf install gh -y
 
-# getting latest Terraform and Packer installers from iac.sh
-## see https://github.com/robertpeteuil/terraform-installer and https://github.com/robertpeteuil/packer-installer
-curl iac.sh/terraform | bash && ./terraform-install.sh -a
-curl iac.sh/packer | bash && ./packer-install.sh -a
+# getting latest HashiCorp tooling (Packer, Terraform, Consul, Vault, Nomad)
+git clone https://github.com/kral2/hashistack-installer/
+cd hashistack-installer || exit
+./hashistack-install.sh
+cd /root || exit
 
 # Python tooling
 ## update setuptools, wheel and pip