+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Bootstrap.sh installs new tools:
 
 - project_o : a smart oci-cli wrapper
 - GitHub cli
+- full HashiCorp stack: Packer, Terraform, Consul, Vault, Nomad
 
 ### Changed